+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.0.14 (2022-11-08)
+
 ### Added
 
 -   `GetFormSubmissionAttachmentMeta` to `FormsClient`